requirements.txt
pathway[llm]
google-generativeai
sentence-transformers
python-dotenv
pandas
requests